+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/tasks/SentryUploadNativeSymbolsTask.kt
@@ -10,54 +10,21 @@ import io.sentry.android.gradle.util.hookWithAssembleTasks
 import io.sentry.android.gradle.util.info
 import io.sentry.gradle.common.SentryVariant
 import java.io.File
-import org.apache.tools.ant.taskdefs.condition.Os
-import org.apache.tools.ant.taskdefs.condition.Os.FAMILY_WINDOWS
 import org.gradle.api.Project
-import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
-import org.gradle.api.tasks.Exec
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.Internal
-import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.TaskProvider
 
-abstract class SentryUploadNativeSymbolsTask : Exec() {
+abstract class SentryUploadNativeSymbolsTask : SentryCliExecTask() {
 
     init {
         description = "Uploads native symbols to Sentry"
     }
 
     @get:Input
-    @get:Optional
-    abstract val debug: Property<Boolean>
-
-    @get:Input
-    abstract val cliExecutable: Property<String>
-
-    @get:Input
     abstract val autoUploadNativeSymbol: Property<Boolean>
-
-    @get:InputFile
-    @get:Optional
-    abstract val sentryProperties: RegularFileProperty
-
-    @get:Input
-    @get:Optional
-    abstract val sentryOrganization: Property<String>
-
-    @get:Input
-    @get:Optional
-    abstract val sentryProject: Property<String>
-
-    @get:Input
-    @get:Optional
-    abstract val sentryAuthToken: Property<String>
-
-    @get:Input
-    @get:Optional
-    abstract val sentryUrl: Property<String>
 
     @get:Input
     abstract val includeNativeSources: Property<Boolean>
@@ -65,68 +32,12 @@ abstract class SentryUploadNativeSymbolsTask : Exec() {
     @get:Internal
     abstract val variantName: Property<String>
 
-    @get:Internal
-    abstract val sentryTelemetryService: Property<SentryTelemetryService>
-
-    override fun exec() {
-        computeCommandLineArgs().let {
-            commandLine(it)
-            logger.info { "cli args: $it" }
-        }
-        setSentryPropertiesEnv()
-        setSentryAuthTokenEnv()
-        super.exec()
-    }
-
-    internal fun setSentryPropertiesEnv() {
-        val sentryProperties = sentryProperties.orNull
-        if (sentryProperties != null) {
-            environment("SENTRY_PROPERTIES", sentryProperties)
-        } else {
-            logger.info { "sentryProperties is null" }
-        }
-    }
-
-    internal fun setSentryAuthTokenEnv() {
-        val sentryAuthToken = sentryAuthToken.orNull
-        if (sentryAuthToken != null) {
-            environment("SENTRY_AUTH_TOKEN", sentryAuthToken)
-        } else {
-            logger.info { "sentryAuthToken is null" }
-        }
-    }
-
-    internal fun computeCommandLineArgs(): List<String> {
-        val args = mutableListOf(
-            cliExecutable.get()
-        )
-
-        sentryUrl.orNull?.let {
-            args.add("--url")
-            args.add(it)
-        }
-
+    override fun getArguments(args: MutableList<String>) {
         args.add("debug-files")
         args.add("upload")
 
-        if (debug.getOrElse(false)) {
-            args.add("--log-level=debug")
-        }
-
-        sentryTelemetryService.orNull?.traceCli()?.let { args.addAll(it) }
-
         if (!autoUploadNativeSymbol.get()) {
             args.add("--no-upload")
-        }
-
-        sentryOrganization.orNull?.let {
-            args.add("--org")
-            args.add(it)
-        }
-
-        sentryProject.orNull?.let {
-            args.add("--project")
-            args.add(it)
         }
 
         val sep = File.separator
@@ -144,12 +55,6 @@ abstract class SentryUploadNativeSymbolsTask : Exec() {
         if (includeNativeSources.get()) {
             args.add("--include-sources")
         }
-
-        if (Os.isFamily(FAMILY_WINDOWS)) {
-            args.add(0, "cmd")
-            args.add(1, "/c")
-        }
-        return args
     }
 
     companion object {