+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryCliProviderTest.kt
@@ -2,9 +2,9 @@ package io.sentry.android.gradle
 
 import io.sentry.android.gradle.SentryCliProvider.extractCliFromResources
 import io.sentry.android.gradle.SentryCliProvider.getCliSuffix
+import io.sentry.android.gradle.SentryCliProvider.getResourceUrl
 import io.sentry.android.gradle.SentryCliProvider.getSentryPropertiesPath
 import io.sentry.android.gradle.SentryCliProvider.searchCliInPropertiesFile
-import io.sentry.android.gradle.SentryCliProvider.searchCliInResources
 import io.sentry.android.gradle.util.SystemPropertyRule
 import io.sentry.android.gradle.util.WithSystemProperty
 import java.io.File
@@ -130,7 +130,7 @@ class SentryCliProviderTest {
                 createNewFile()
             }
 
-        val foundPath = searchCliInResources(resourcePath)
+        val foundPath = getResourceUrl(resourcePath)
         assertNotNull(foundPath)
         assertTrue(
             foundPath.endsWith("${File.separator}dummy-bin${File.separator}dummy-sentry-cli")
@@ -143,7 +143,7 @@ class SentryCliProviderTest {
     fun `searchCliInResources returns null if file does not exist`() {
         val resourcePath = "./dummy-bin/i-dont-exist"
 
-        assertNull(searchCliInResources(resourcePath))
+        assertNull(getResourceUrl(resourcePath))
     }
 
     @Test
@@ -157,20 +157,22 @@ class SentryCliProviderTest {
                 writeText("echo \"This is just a dummy script\"")
             }
 
-        val loadedPath = extractCliFromResources(File("."), resourcePath)
+        val outputFile = File("bin", "output-bin")
+        val loadedPath = extractCliFromResources(resourcePath, outputFile)
         assertNotNull(loadedPath)
 
         val binContent = File(loadedPath).readText()
         assertEquals("echo \"This is just a dummy script\"", binContent)
 
         resourceFile?.delete()
+        outputFile.delete()
     }
 
     @Test
     fun `loadCliFromResourcesToTemp returns null if file does not exist`() {
         val resourcePath = "./dummy-bin/i-dont-exist"
 
-        assertNull(extractCliFromResources(File("."), resourcePath))
+        assertNull(extractCliFromResources(resourcePath, File(".")))
     }
 
     @Test